requirements.txt
streamlit==1.28.1
plotly==5.15.0
pandas==2.0.3
google-api-python-client==2.104.0
python-dotenv==1.0.0